[Core] Relocate session memory under experimental/world_models/
Move the RFC #4480 Phase 0 session memory package from vllm_omni/diffusion/memory/ to vllm_omni/experimental/world_models/memory/, and the DreamZero adapter to vllm_omni/experimental/world_models/adapters/, per review feedback that world-model session memory belongs under experimental/. Imports updated; no behavior change.
