+++ vllm_omni/diffusion/models/dreamzero/pipeline_dreamzero.py
@@ -36,12 +36,10 @@ from vllm_omni.diffusion.distributed.autoencoders.autoencoder_kl_wan import (
 from vllm_omni.diffusion.distributed.cfg_parallel import CFGParallelMixin
 from vllm_omni.diffusion.distributed.parallel_state import get_classifier_free_guidance_world_size
 from vllm_omni.diffusion.distributed.utils import get_local_device
-from vllm_omni.diffusion.memory import SessionMemoryManager, resolve_session_memory_config
 from vllm_omni.diffusion.model_loader.diffusers_loader import DiffusersPipelineLoader
 from vllm_omni.diffusion.models.dreamzero.causal_wan_model import CausalWanModel
 from vllm_omni.diffusion.models.dreamzero.image_encoder import DreamZeroImageEncoder
 from vllm_omni.diffusion.models.dreamzero.state_dreamzero import DreamZeroState
-from vllm_omni.diffusion.models.dreamzero.state_dreamzero_adapter import DreamZeroStateAdapter
 from vllm_omni.diffusion.models.dreamzero.transform import (
     DEFAULT_EMBODIMENT,
     ensure_transforms_loaded,
@@ -62,6 +60,11 @@ from vllm_omni.experimental.ar_diffusion.capability import (
     ARDiffusionCrossAttentionKVSpec,
     ARDiffusionKVBranchSpec,
     ARDiffusionKVCacheSpec,
+)
+from vllm_omni.experimental.world_models.adapters.state_dreamzero_adapter import DreamZeroStateAdapter
+from vllm_omni.experimental.world_models.memory import (
+    SessionMemoryManager,
+    resolve_session_memory_config,
 )
 from vllm_omni.inputs.data import OmniDiffusionSamplingParams
