+++ vllm_omni/experimental/world_models/adapters/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+"""Per-model adapters onto the world-model session memory contract (RFC #4480)."""
+
+from __future__ import annotations
